% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{distributionPeak}
\alias{distributionPeak}
\title{Choose samples from the posterior and find index P peaks by years.}
\usage{
distributionPeak(nSamples = NA, CIlevel = 0.95)
}
\arguments{
\item{nSamples}{An integer defining the number of samples to apply the function on. Default is all the samples in the input.}

\item{CIlevel}{A numeric value between 0 and 1, indicating the level of confidence for generated confidence intervals (CI).}
}
\value{
Returns a list with the max peak estimation, a low CI bound for the peak time, and
        a high CI bound for the peak time. This list is also saved as a csv file. Additionally, it produces a pdf barplot
        of peak times colored by year.
}
\description{
Choose samples from the posterior and find index P peaks by years.
}
