% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{estimateEcoCoefficients}
\alias{estimateEcoCoefficients}
\title{Estimates the factors Alpha, Eta and Rho.}
\usage{
estimateEcoCoefficients(
  nMCMC = 30000,
  bMCMC = 0.5,
  cRho = 0.5,
  cEta = 2,
  gauJump = 0.05
)
}
\arguments{
\item{nMCMC}{Number of MCMC steps to run the routine for.}

\item{bMCMC}{Number 0-1 representing the proportion of last steps to be considered3
not burn-in. E.g. 0.7 will use the last 0.3 (30\%) of MCMC steps for posteriors.}

\item{cRho}{Initial guess for factor Rho.}

\item{cEta}{Initial guess for factor Eta.}

\item{gauJump}{Gaussian standard deviation for Rho and Eta jumps in the MCMC.}
}
\description{
For the estimation of Eta and Rho
an MCMC routine is run (for which parameters are available). The function sets
the \code{MVSE_results} global variable which can be used by other functions.
To see a list of globally accessible variables use \code{help_globalVariables()}.
}
