% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{expectedPosterior}
\alias{expectedPosterior}
\title{Choose samples from the posterior and fit filter.}
\usage{
expectedPosterior(nSamples = NA, breakyears = FALSE, CIlevel = 0.95)
}
\arguments{
\item{nSamples}{An integer defining the number of samples to apply filtering to. Default is all the samples in the input.}

\item{breakyears}{A logical variable indicating weather the data should be broken to years and analyzed separately.}

\item{CIlevel}{A numeric value between 0 and 1, indicating the level of confidence for generated confidence intervals (CI).}
}
\value{
Returns a list with the average filter prediction, CI bounds of the average filter values,
        CI bounds for all filter values, and average index P values. This list is also saved as csv files.
        Additionally, ir produces a pdf plot of the filter with overlayed average index P.
}
\description{
Choose samples from the posterior and fit filter.
}
