% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{exportEcoCoefficients}
\alias{exportEcoCoefficients}
\title{Exports the estimated factors Alpha, Rho and Eta into csv files. Four files
are exported. One summarises the means and standard deviations of the estimated
distributions. The other 3, one for each factor, exports all estimated values
that compose the distributions.}
\usage{
exportEcoCoefficients(sep = ",")
}
\arguments{
\item{sep}{CSV field separator character.}
}
\description{
Exports the estimated factors Alpha, Rho and Eta into csv files. Four files
are exported. One summarises the means and standard deviations of the estimated
distributions. The other 3, one for each factor, exports all estimated values
that compose the distributions.
}
