% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{exportEmpiricalIndexP}
\alias{exportEmpiricalIndexP}
\title{Exports the index P estimations into a csv file. Includes mean P and 95\% CI.}
\usage{
exportEmpiricalIndexP(sep = ",")
}
\arguments{
\item{sep}{CSV field separator character.}
}
\description{
Exports the index P estimations into a csv file. Includes mean P and 95\% CI.
}
