% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{exportEmpiricalQ}
\alias{exportEmpiricalQ}
\title{Exports the factor Q estimations into a csv file. Includes mean Q and 95\% CI.}
\usage{
exportEmpiricalQ(sep = ",")
}
\arguments{
\item{sep}{CSV field separator character.}
}
\description{
Exports the factor Q estimations into a csv file. Includes mean Q and 95\% CI.
}
