% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{exportEmpiricalV0}
\alias{exportEmpiricalV0}
\title{Exports the V0 estimations into a csv file. Includes mean V0 and 95\% CI.}
\usage{
exportEmpiricalV0(sep = ",")
}
\arguments{
\item{sep}{CSV field separator character.}
}
\description{
Exports the V0 estimations into a csv file. Includes mean V0 and 95\% CI.
}
