% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{exportEntoParameters}
\alias{exportEntoParameters}
\title{Exports the estimated entomological parameters of mosquito life-span, mosquito
incubation period and mosquito biting rate into one single CSV file.}
\usage{
exportEntoParameters(sep = ",", Ns = 200)
}
\arguments{
\item{sep}{CSV field separator character.}

\item{Ns}{number of samples wanted.}
}
\description{
Exports the estimated entomological parameters of mosquito life-span, mosquito
incubation period and mosquito biting rate into one single CSV file.
}
