% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{exportGenerationTimes}
\alias{exportGenerationTimes}
\title{Exports the estimated generation times.}
\usage{
exportGenerationTimes(sep = ",")
}
\arguments{
\item{sep}{CSV field separator character.}
}
\description{
Exports the estimated generation times.
}
