% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{getClimateInputFile}
\alias{getClimateInputFile}
\title{Get the file name and path of the climatic series input.}
\usage{
getClimateInputFile()
}
\description{
Get the file name and path of the climatic series input.
}
