% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{getMonthlyIndexP}
\alias{getMonthlyIndexP}
\title{Calculates the mean monthly index P.}
\usage{
getMonthlyIndexP()
}
\value{
data.frame with colums mean index P (meanP), stdev index P (sdP), date
in the form 'year-month' (YM).
}
\description{
Calculates the mean monthly index P.
}
