% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{getPosteriorGammaVForTimet}
\alias{getPosteriorGammaVForTimet}
\title{Gets the posterior of the vector incubation period. User can choose particular
time step t and use mEtaRho samples to calculate the incubation period from the
posterior of the ecological coefficient alpha.
If the number of samples is not given, it is assumed the user wants a 'mean'
estimation, and the mean of the alpha ecological coefficient is used over all
time steps.}
\usage{
getPosteriorGammaVForTimet(mAlpha = NA, t = NA)
}
\arguments{
\item{mAlpha}{if given and numeric, sets the number of samples to be taken from
from the alpha ecological parameter that is used to calculate the incubation
period.}

\item{t}{if given and numeric, sets the desired time step to get a distribution
of the incubation period on that step.}
}
\value{
incubation period.
}
\description{
Gets the posterior of the vector incubation period. User can choose particular
time step t and use mEtaRho samples to calculate the incubation period from the
posterior of the ecological coefficient alpha.
If the number of samples is not given, it is assumed the user wants a 'mean'
estimation, and the mean of the alpha ecological coefficient is used over all
time steps.
}
