% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{getPosteriorLifespanVForTimet}
\alias{getPosteriorLifespanVForTimet}
\title{Gets the posterior of the vector life span. User can choose particular
time step t and use mEtaRho samples to calculate the life span from the posteriors
of the ecological coefficients eta and rho.
If the number of samples is not given, it is assumed the user wants a 'mean'
estimation, and the mean of the alpha ecological coefficient is used over all
time steps.}
\usage{
getPosteriorLifespanVForTimet(mEtaRho = NA, t = NA)
}
\arguments{
\item{mEtaRho}{if given and numeric, sets the number of samples to be taken from
from the eta and rho ecological parameters that are used to calculate the
life span.}

\item{t}{if given and numeric, sets the desired time step to get a distribution
of the life span on that step.}
}
\value{
incubation period.
}
\description{
Gets the posterior of the vector life span. User can choose particular
time step t and use mEtaRho samples to calculate the life span from the posteriors
of the ecological coefficients eta and rho.
If the number of samples is not given, it is assumed the user wants a 'mean'
estimation, and the mean of the alpha ecological coefficient is used over all
time steps.
}
