% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{getTempEffOviPos}
\alias{getTempEffOviPos}
\title{Gets the time series for the effect of temperature on oviposition. Please see
Lourenco et al. doi.org/10.7554/eLife.29820.001 for details on how this effect
is calculated.}
\usage{
getTempEffOviPos(tpe = "emp", t = NA)
}
\arguments{
\item{tpe}{defines the type of query: 'emp' (default) for empirical range; any other
value will return the theoretical range (used in the index P suitability maps).}

\item{t}{should be numeric and if not NA can be used to obtain the value at a
particular time step t.}
}
\value{
effect as array.
}
\description{
Gets the time series for the effect of temperature on oviposition. Please see
Lourenco et al. doi.org/10.7554/eLife.29820.001 for details on how this effect
is calculated.
}
