% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{getTypicalYear}
\alias{getTypicalYear}
\title{Calculates the mean behaviour of a year using the climatic data and the estimated P
for several years (if the input is for several years).}
\usage{
getTypicalYear()
}
\value{
data.frame with colums mean temperature (meanT), stdev temperature (sdT),
mean humidity (meanH), stdev humidity (sdH), mean index P (meanP), stdev index P (sdP)
}
\description{
Calculates the mean behaviour of a year using the climatic data and the estimated P
for several years (if the input is for several years).
}
