% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{help_inputFormat}
\alias{help_inputFormat}
\title{Shows detailed information on the required csv format for the climatic input
data series.}
\usage{
help_inputFormat()
}
\description{
Shows detailed information on the required csv format for the climatic input
data series.
}
