% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsefit.R
\name{mcmc_pairs}
\alias{mcmc_pairs}
\alias{mcmc_pairs,mvsefit-method}
\title{Plot a pairs plot of the MCMC draws of the factors rho, eta and alpha}
\usage{
\S4method{mcmc_pairs}{mvsefit}(object, filename = NULL, options = list(max_draws = 10^4))
}
\arguments{
\item{object}{An object of S4 class \code{\linkS4class{mvsefit}}.}

\item{filename}{The name of the file where the plot in PNG format is saved (optional)..}

\item{options}{A named list of options for the plot.}
}
\value{
Many ggplot objects organized into a grid via \code{\link{bayesplot_grid}} is returned.
}
\description{
A square plot matrix with univariate marginal distributions along the diagonal (as histograms) and
bivariate distributions off the diagonal (as scatterplots).
}
\examples{
# obtain climate data
data(climateFSA)

# define a mvse model 
priors <- list()
priors$mosq_life_exp <- list(dist="normal", pars=c(mean=12, sd=2)) # mosquito life expectancy (days)
priors$mosq_inc_per <- list(dist="normal", pars=c(mean=7, sd=2))   # mosquito-virus incubation period (days)
priors$mosq_biting_freq <- list(dist="normal", pars=c(mean=0.25, 0.01)) # mosquito biting frequency (bites/female/day)
priors$human_life_exp <- list(dist="normal", pars=c(mean=71.1, sd=2)) # human life expectancy (years)
priors$human_inc_per <- list(dist="normal", pars=c(mean=5.8, sd=1)) # human-virus incubation period (days)
priors$human_inf_per <- list(dist="normal", pars=c(mean=5.9, sd=1)) # human-virus infectious period (days)
user_model <- mvse_model(model_name="my_mvse_model", climate_data=climateFSA, priors=priors)

# run the MCMC sampling procedure to estimate the epi-entomological parameters as well Index P, Q and V0
user_fit <- fitting(object=user_model, iter=10^5, warmup=0.2, seed=123, init=c(rho=1, eta=10, alpha=3), samples=1000)

# plot the pairs plot for the MCMC draws of alpha, rho, eta
mcmc_pairs(user_fit, options=list(max_draws=10^5))

}
