% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsefit.R
\name{mcmc_traceplot}
\alias{mcmc_traceplot}
\alias{mcmc_traceplot,mvsefit-method}
\title{Plot trace plots of the MCMC draws of the factors rho, eta and alpha}
\usage{
\S4method{mcmc_traceplot}{mvsefit}(object, filename = NULL, options = list(max_iter = 10^4))
}
\arguments{
\item{object}{An object of S4 class \code{\linkS4class{mvsefit}}.}

\item{filename}{The name of the file where the plot in PNG format is saved (optional).}

\item{options}{A named list of options for the plot.}
}
\value{
Many ggplot objects organized into a grid via \code{\link{bayesplot_grid}} is returned.
}
\description{
Displays a separate plot of iterations vs. sampled values for each of the factors.
}
\examples{
# obtain climate data
data(climateFSA)

# define a mvse model 
aegypti_model <- mvse_model(model_name="my_aegypti_mvse_model", model_category="aegypti", climate_data=climateFSA)

# run the MCMC sampling procedure to estimate the epi-entomological parameters as well Index P, Q and V0
aegypti_fit <- fitting(object=aegypti_model, iter=10^5, warmup=0.2, seed=123, init=c(rho=1, eta=10, alpha=3), samples=1000)

mcmc_traceplot(aegypti_fit, options=list(max_iter=10^5))
}
