% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE-package.R
\docType{package}
\name{MVSE-package}
\alias{MVSE}
\alias{MVSE-package}
\title{MVSE: A package for estimating (M)osquito-borne (V)iral (S)uitability (E)stimator}
\description{
The MVSE package provides computational methods to estimate an index of 'transmission potential' for mosquito-borne viruses. MVSE allows for
parameterization of particular viruses, host and mosquito-species. The methods
offered are climate-driven and can therefore be location specific.
}
\details{
For more information, please refer to the manual.
}
\author{
Taishi Nakase
}
\keyword{internal}
