% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsefit.R
\docType{class}
\name{mvsefit-class}
\alias{mvsefit-class}
\title{Class \code{mvsefit}: fitted MVSE model}
\description{
An object of class \code{mvsefit} contains the output derived from sampling of a
MVSE model as returned by the method \link{sampling}.
}
\section{Slots}{

\describe{
\item{\code{model_name}}{The model name as a string.}

\item{\code{model_category}}{The model category as a string.}

\item{\code{mvsemodel}}{The instance of S4 class \code{\linkS4class{mvsemodel}}.}

\item{\code{mvse_args}}{A list containing the arguments used for sampling (e.g. \code{iter}, \code{warmup}, \code{init}, \code{gauJumps}, etc.).}

\item{\code{sim}}{A list containing simulation results including the posterior draws for the factors (rho, eta, alpha) and the generated
quantities (i.e. indexP) as well as various pieces of data used by many of the methods for \code{mvsefit} objects.}
}}

\section{Methods}{

\subsection{Printing, plotting and summarizing:}{
\describe{\item{\code{show}}{Print the default summary for the model.}}}
\subsection{Extracting posterior draws:}{
\describe{\item{\code{extract}}{Extact the sampled posterior distribution of the Index P as well as the ecological scaling parameters.}}}
\subsection{Simulating:}{
\describe{\item{\code{simulate}}{Simulate Index P using the fitted model.}}}
}

