% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsemodel.R
\docType{class}
\name{mvsemodel-class}
\alias{mvsemodel-class}
\title{Class representing MVSE model}
\description{
A \code{mvsemodel} object contains both climate data and prior distributions
for  human/vector/virus biological parameters. The \code{sampling}
method defined in this class may be used to draw samples of the ecological scaling coefficients
(rho, eta, alpha) from this model.
}
\details{
Instances of \code{mvsemodel} are created by calling the function \code{\link{mvse_model}}.
}
\section{Slots}{

\describe{
\item{\code{model_name}}{The model name, an object of type \code{character}.}

\item{\code{model_category}}{The model category, an object of type \code{character}.}

\item{\code{climate_data}}{The climate data, an object of type \code{dataframe}.}

\item{\code{priors}}{The prior distributions of the human/vector/virus biological parameters,
an object of type \code{list}.}
}}

\section{Methods}{

\describe{\item{\code{show}}{Print the default summary for the model.}}
\describe{\item{\code{print}}{Print all information contained in the model.}}
}

