% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsemodel.R
\name{plot_climate}
\alias{plot_climate}
\alias{plot_climate,mvsemodel-method}
\title{Plot climate data for \code{mvsemodel} objects}
\usage{
\S4method{plot_climate}{mvsemodel}(
  object,
  vars = c("temperature", "humidity"),
  smoothing = NULL,
  filename = NULL
)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{mvsemodel}}.}

\item{vars}{A vector of strings specifying the climatic variables to plot. Possible arguments include
\code{"temperature"}, \code{"humidity"}, or the pair of the two strings.}

\item{smoothing}{An optional positive integer used for smoothing the climate data. For example,
7 will smooth the time series using +- 7 time points per existing point (up and down that point).}

\item{filename}{The name of the file where the plot in PNG format is saved (optional).}
}
\value{
A ggplot object that can be further customized using the \bold{ggplot2} package.
}
\description{
The default plot shows the time series for temperature and humidity. The method can also be used to
plot other combinations of climatic variables (i.e. temperature, humidity and rainfall).
}
\examples{
#define a mvse model
data(climateFSA)
user_model <- mvse_model(model_name="my_mvse_model", climate_data=climateFSA, priors=list())

# plot climate data
plot_climate(object=user_model) # temperature and humidity
plot_climate(object=user_model, vars=c("temperature"))
plot_climate(object=user_model, vars=c("temperature"), filename="climate_plot.png")

}
\seealso{
\code{\linkS4class{mvsemodel}}
}
