% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsemodel.R
\name{plot_priors}
\alias{plot_priors}
\alias{plot_priors,mvsemodel-method}
\title{Plot the prior distributions of the host/vector ecological/epidemiological parameters.}
\usage{
\S4method{plot_priors}{mvsemodel}(object, pars = NULL, filename = NULL)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{mvsemodel}}.}

\item{pars}{If not NULL, a character vector indicating which parameters to include in the plots. By
default, all parameters are included.}

\item{filename}{The name of the file where the plot in PNG format is saved (optional).}
}
\value{
If only one parameter is specified, a single ggplot object is returned. Otherwise, many ggplot
objects organized into a grid via \code{\link{bayesplot_grid}} is returned.
}
\description{
The default plot shows the probability density function for all parameters.
}
\examples{
#define a mvse model
data(climateFSA)
priors <- list()
priors$mosq_life_exp <- list(dist="normal", pars=c(mean=12, sd=2)) # mosquito life expectancy (days)
priors$mosq_inc_per <- list(dist="normal", pars=c(mean=7, sd=2))   # mosquito-virus incubation period (days)
priors$mosq_biting_freq <- list(dist="normal", pars=c(mean=0.25, 0.01)) # mosquito biting frequency (bites/female/day)
priors$human_life_exp <- list(dist="normal", pars=c(mean=71.1, sd=2)) # human life expectancy (years)
priors$human_inc_per <- list(dist="normal", pars=c(mean=5.8, sd=1)) # human-virus incubation period (days)
priors$human_inf_per <- list(dist="normal", pars=c(mean=5.9, sd=1)) # human-virus infectious period (days)
user_model <- mvse_model(model_name="my_mvse_model", climate_data=climateFSA, priors=list())

# plot prior distributions of epi-entomological parameters
plot_priors(object=user_model)
plot_priors(object=user_model, pars=c("mosq_life_exp", "mosq_inc_per", "human_inf_per"))

}
\seealso{
\code{\linkS4class{mvsemodel}}
}
