% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotClimate}
\alias{plotClimate}
\title{Plots and exports a simple figure (mostly) for debug purposes. The mean index
P and 95\% CI are plotted.}
\usage{
plotClimate(outfilename = "debug_climate")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports a simple figure (mostly) for debug purposes. The mean index
P and 95\% CI are plotted.
}
