% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEcoCoefMCMCChains}
\alias{plotEcoCoefMCMCChains}
\title{Plots and exports a simple figure (mostly) for debug purposes. The MCMC chains
are presented for the factors Eta and Rho.}
\usage{
plotEcoCoefMCMCChains(outfilename = "debug_chains")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports a simple figure (mostly) for debug purposes. The MCMC chains
are presented for the factors Eta and Rho.
}
