% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEcoCoefPosteriors}
\alias{plotEcoCoefPosteriors}
\title{Plots and exports a simple figure (mostly) for debug purposes. The distributions
of factors Alpha, Eta and Rho are plotted using R's default \code{density()}
function.}
\usage{
plotEcoCoefPosteriors(
  outfilename = "debug_dist",
  etaLim = c(),
  alphaLim = c(),
  rhoLim = c()
)
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports a simple figure (mostly) for debug purposes. The distributions
of factors Alpha, Eta and Rho are plotted using R's default \code{density()}
function.
}
