% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalIndexP}
\alias{plotEmpiricalIndexP}
\title{Plots and exports the mean index
P and 95\% CI under smoothing and no smoothing.}
\usage{
plotEmpiricalIndexP(outfilename = "debug_indexP")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports the mean index
P and 95\% CI under smoothing and no smoothing.
}
