% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalIndexPPV0PQ}
\alias{plotEmpiricalIndexPPV0PQ}
\title{Plots and exports the mean index P, Q*P and V0*P}
\usage{
plotEmpiricalIndexPPV0PQ(outfilename = "debug_mean_P_PV0_PQ")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports the mean index P, Q*P and V0*P
}
