% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalIndexPV0Q}
\alias{plotEmpiricalIndexPV0Q}
\title{Plots and exports the mean index P, Q and V0}
\usage{
plotEmpiricalIndexPV0Q(outfilename = "debug_mean_P_Q_V0")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports the mean index P, Q and V0
}
