% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalQ}
\alias{plotEmpiricalQ}
\title{Plots and exports the mean Q
and 95\% CI under smoothing and no smoothing.}
\usage{
plotEmpiricalQ(outfilename = "debug_Q")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports the mean Q
and 95\% CI under smoothing and no smoothing.
}
