% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalSuitRespMap}
\alias{plotEmpiricalSuitRespMap}
\title{Plots the response, in terms of empirical index P, into a bubble chart with climated
as input variables and months identified in the range of the climatic vars.}
\usage{
plotEmpiricalSuitRespMap(
  outfilename = "debug_empirical_response",
  MinVarNorm = NA,
  MaxVarNorm = NA,
  TempLim = c(),
  HumLim = c(),
  cols = rev(c("red3", "gold", "skyblue", "violet", "purple"))
)
}
\arguments{
\item{outfilename}{A name tag for the output file.}

\item{MinVarNorm}{scale colours to minimum desired value}

\item{MaxVarNorm}{scale colours to maximum desired value}

\item{TempLim}{Two numbers for the limit of temperature to be plotted.}

\item{HumLim}{Two numbers for the limit of humidity to be plotted.}

\item{cols}{Colours to be used in the colour key.}
}
\description{
Plots the response, in terms of empirical index P, into a bubble chart with climated
as input variables and months identified in the range of the climatic vars.
}
