% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalV0}
\alias{plotEmpiricalV0}
\title{Plots and exports the mean V0
and 95\% CI under smoothing and no smoothing.}
\usage{
plotEmpiricalV0(outfilename = "debug_V0")
}
\arguments{
\item{outfilename}{A name tag for the output file.}
}
\description{
Plots and exports the mean V0
and 95\% CI under smoothing and no smoothing.
}
