% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEmpiricalVecCapMap}
\alias{plotEmpiricalVecCapMap}
\title{Plots the response, in terms of empirical capacity, into a bubble chart with climated
as input variables and months identified in the range of the climatic vars.}
\usage{
plotEmpiricalVecCapMap(
  outfilename = "debug_empirical_vecCap",
  MinVarNorm = NA,
  MaxVarNorm = NA,
  TempLim = c(),
  HumLim = c(),
  cols = rev(c("red3", "gold", "skyblue", "violet", "purple"))
)
}
\arguments{
\item{outfilename}{A name tag for the output file.}

\item{MinVarNorm}{scale colours to minimum desired value}

\item{MaxVarNorm}{scale colours to maximum desired value}

\item{TempLim}{Two numbers for the limit of temperature to be plotted.}

\item{HumLim}{Two numbers for the limit of humidity to be plotted.}

\item{cols}{Colours to be used in the colour key.}
}
\description{
Capacity is the expected proportion of human to mosquito to human transmission events
that would have mosquito lifespan longer than the mosquito incubation period. Capacity
towards zero implies unlikeliness that the mosquitoes will live longer enough to transmit.
}
