% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotEntoParameters}
\alias{plotEntoParameters}
\title{Plots and exports a simple figure (mostly) for debug purposes. The entomological
parameters that depend on the factors Alpha, Eta and Rho are plotted. These
parameters are the mosquito life-span (1/death rate), the mosquito incubation
period (1/rate incubation to infection), mosquito biting rate. Both the
resulting time series and distributions are presented. The distributions should
mimic the priors used before estimating the factors Alpha, Eta and Rho. The
priors are simulated when plotted and also presented.}
\usage{
plotEntoParameters(
  outfilename = "debug_ento",
  Ns = 200,
  entoPostLim = c(),
  bitPostLim = c()
)
}
\arguments{
\item{outfilename}{A name tag for the output file.}

\item{Ns}{Number of samples wanted for posteriors.}

\item{entoPostLim}{Two numbers for the y limits to be used in the posterior plots.}

\item{bitPostLim}{Two numbers for the y limits to be used in the posterior plots.}
}
\description{
Plots and exports a simple figure (mostly) for debug purposes. The entomological
parameters that depend on the factors Alpha, Eta and Rho are plotted. These
parameters are the mosquito life-span (1/death rate), the mosquito incubation
period (1/rate incubation to infection), mosquito biting rate. Both the
resulting time series and distributions are presented. The distributions should
mimic the priors used before estimating the factors Alpha, Eta and Rho. The
priors are simulated when plotted and also presented.
}
