% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{plotGenerationTimes}
\alias{plotGenerationTimes}
\title{Plots and exports generation time and variables related to its calculation.}
\usage{
plotGenerationTimes(
  outfilename = "debug_genTimes",
  entoLim = c(),
  humLim = c(),
  genLim = c()
)
}
\arguments{
\item{outfilename}{A name tag for the output file.}

\item{entoLim}{two numbers setting the y-axis limits for the entomological solutions}

\item{humLim}{two numbers setting the y-axis limits for the human solutions}

\item{genLim}{two numbers setting the y-axis limits for the generation solutions}
}
\description{
Plots and exports generation time and variables related to its calculation.
}
