% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsefit.R
\name{print,mvsefit-method}
\alias{print,mvsefit-method}
\title{Print a customizable summary for the model.}
\usage{
\S4method{print}{mvsefit}(x, pars = NULL, probs = c(0.25, 0.5, 0.75), digits_summary = 3, ...)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{mvsefit}}.}

\item{pars}{A character vector of parameter names. The default is all parameters
(i.e. \code{"rho"}, \code{"eta"}, \code{"alpha"}, \code{"indexP"}).}

\item{probs}{A numeric vector of quantiles of interest. The default is
\code{c(0.25, 0.5, 0.75)}.}

\item{digits_summary}{The number of decimal places to use when printing the summary, defaulting to 3.}

\item{...}{Additional arguments passed to the \code{print} method for \code{mvsefit} objects.}
}
\description{
Print a customizable summary for the model.
}
