% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsemodel.R
\name{print,mvsemodel-method}
\alias{print,mvsemodel-method}
\title{Print the default summary of model (with all information).}
\usage{
\S4method{print}{mvsemodel}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{mvsemodel}}.}

\item{...}{Additional arguments passed to the \code{print} method for \code{mvsemodel} objects.}
}
\description{
Print the default summary of model (with all information).
}
