% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{setEmpiricalClimateSeries}
\alias{setEmpiricalClimateSeries}
\title{Reads the climatic series and applies some required transformations ('digests').
For required formats, please see information in \code{help_inputFormat()}.}
\usage{
setEmpiricalClimateSeries(filepath = NA, D = NA, sep = ",", NDTSmoothTH = 30)
}
\arguments{
\item{filepath}{File path to file where the data to be used is.}

\item{D}{data.frame with the data to use.}

\item{sep}{Character separator for csv file.}

\item{NDTSmoothTH}{Number of time points to smooth the temperature and humidity with.}
}
\description{
The effects of climateic variables on entomological parameters are calculated here.
These include the effects of temperature on mosquito lifespan and probability of
transmission per mosquito bite, and humidity on lifespan and biting rate.
Please see Lourenco et al. doi.org/10.7554/eLife.29820.001 for details on how
these effects are calculated.
}
\details{
The function sets a series of global variables which can be
used by other functions. To see a list of globally accessible variables use
\code{help_globalVariables()}.
}
