% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{setHumanLifeExpPrior}
\alias{setHumanLifeExpPrior}
\title{Set the prior for the human life-span.}
\usage{
setHumanLifeExpPrior(pmean, psd, pdist)
}
\arguments{
\item{pmean}{A number.}

\item{psd}{A number.}

\item{pdist}{A string naming the desired3 distribution: implemented 'gaussian'
'exponential' and 'gamma'.}
}
\description{
Set the prior for the human life-span.
}
