% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{setHumanMosqTransProbPrior}
\alias{setHumanMosqTransProbPrior}
\title{Set the prior human to mosquito probability of transmission per infectious bite}
\usage{
setHumanMosqTransProbPrior(pmean, psd, pdist)
}
\arguments{
\item{pmean}{A number.}

\item{psd}{A number.}

\item{pdist}{A string naming the desired3 distribution: implemented 'gaussian'
'exponential' and 'gamma'.}
}
\description{
Set the prior human to mosquito probability of transmission per infectious bite
}
