% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{setOutputFilePathAndTag}
\alias{setOutputFilePathAndTag}
\title{Set the file name and path of the climatic series input.}
\usage{
setOutputFilePathAndTag(tag)
}
\arguments{
\item{tag}{A character string.}
}
\description{
Set the file name and path of the climatic series input.
}
