% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{setTheoreticalClimateSeries}
\alias{setTheoreticalClimateSeries}
\title{Reads the climatic series and applies some required transformations ('digests').
For required formats, please see information in \code{help_inputFormat()}.}
\usage{
setTheoreticalClimateSeries(
  Trange = seq(10, 35, length.out = 10),
  Hrange = seq(0, 100, length.out = 10),
  NtimePointsSmooth = 3
)
}
\arguments{
\item{Trange}{Numerical range to be considered3 for temperature in Celsius.}

\item{Hrange}{Numerical range to be considered3 for humidity in the range 0 -100 \%}

\item{NtimePointsSmooth}{Number of time points to smooth the climatic series
with.}
}
\description{
The effects of climateic variables on entomological parameters are calculated here.
These include the effects of temperature on mosquito lifespan and probability of
transmission per mosquito bite, and humidity on lifespan and biting rate.
Please see Lourenco et al. doi.org/10.7554/eLife.29820.001 for details on how
these effects are calculated.
}
\details{
The function sets a series of global variables which can be
used by other functions. To see a list of globally accessible variables use
\code{help_globalVariables()}.
}
