% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsefit.R
\name{show,mvsefit-method}
\alias{show,mvsefit-method}
\title{Print the default summary for the model.}
\usage{
\S4method{show}{mvsefit}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{mvsefit}}.}
}
\description{
Print the default summary for the model.
}
