% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsemodel.R
\name{show,mvsemodel-method}
\alias{show,mvsemodel-method}
\title{Print the default summary for the model}
\usage{
\S4method{show}{mvsemodel}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{mvsemodel}}.}
}
\description{
Print the default summary for the model
}
