% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{simulateEmpiricalIndexP}
\alias{simulateEmpiricalIndexP}
\title{Estimates the index P by simulation. That is, the estimated posteriors of factors
Eta and Rho are sampled, as well as the distribution of factor Alpha. These
independent samples are then used to calculate index P given its mathematical
expression. Since this expression is dependent on climatic variables in time,
the result is a time series of the index P. The user can choose the number of
samples, resulting in \code{nSample} estimations of P in time. These independent
estimations are then used to calculate the mean and 95\% CI of the index P.}
\usage{
simulateEmpiricalIndexP(nSample = 1000, smoothing = c())
}
\arguments{
\item{nSample}{Number of samples to draw from factors Alpha, Eta and Rho.}

\item{smoothing}{Set of integers that define how many averaging smoothers should
be calculated for the index P. E.g. c(7, 30) will calculate index P series using
+- 7 and +- 30 time points per existing point (up and down that point).}
}
\description{
The function adds the resulting index P estimations to the
\code{MVSE_results} global variable which can be used by other functions. Smoothed,
averaged estimations can also be added to \code{MVSE_results} using the parameter
\code{smoothing}. All independent estimations of index P in time are also stored3
in the \code{MVSE_indexP} global variable (matrix, rows as independent estimations,
cols as time steps). To see a list of globally accessible variables use
\code{help_globalVariables()}
}
\details{
The index P follows the expression of R0 from Lourenco et al. doi.org/10.7554/eLife.29820.001
ignoring the number of mosquitoes per human (V/N).
}
