% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{simulateGenerationTime}
\alias{simulateGenerationTime}
\title{Estimates the (one-cycle) generation time; that is, the expected time for one
human case to transmit to a vector, and back to a human. This function uses the
estimated posteriors of vector lifespan and incubation period, together with the
priors for human infectious period and incubation period. The generation time is
separated into 2 time contributions: the time it takes for the human to transmit
to the vector, and the time it takes for the vector to transmit to a second human.
Both of these contributions are simulated and returned to the user. The actual
generation time is the sum of the two contributions.}
\usage{
simulateGenerationTime(Ns = 200)
}
\arguments{
\item{Ns}{Number of samples to be drawn from priors and posteriors per time step.}
}
\description{
It is assumed that transmission may take place at any time point during the
infectious period of the human / vector. That is, while the infectious period
may be X, the 'effective waiting time for transmission' may be shorter. We take
the effective waiting time for transmission of an infectious period X as
\code{runif(1, min=0, max=X)}. The human infectious is taken from the prior;
the vector infectious period is taken from its lifespan minus the incubation
period.
}
\details{
Generation time is calculated independently for every time step t. For a particupar
step, Ns samples are drawn from the human priors of infectious and incubation
periods. Ns samples are also taken for the posteriors of lifespan and incubation
period of the vector. The uniform sampling described above is used for every Ns sample.
The human and vector contributions to the generation times are just the sum of the
incubation period samples and 'effective waiting time for transmission'.

The function adds the resulting estimations to \code{MVSE_results} global
variable under the name 'Tg' (\code{MVSE_results$Tg}). The output is a data.frame
with 22 columns (Tg, Tgu, Tgl, HTg, HTgu, HTgl, VTg, VTgu, VTgl, VLS, VLSu, VLSl,
EIP, EIPu, EIPl, HIP, HIPu, HIPl, IIP, IIPu, IIPl, FrNoVecCap). The first 21
are organized in sets of 3, with mean, 95 CI upper and lower bounds.
Tg stands for generation time, HTg human contribution, VTg vector contribution,
VLS vector lifespan, EIP extrinsic incubation period, HIP human infectious period,
IIP instrinsic incubation period. The last column FrNoVecCap is the proportion of
samples per time step which had vector lifespan longer than the incubation period.
}
