% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{simulateTheoreticalIndexP}
\alias{simulateTheoreticalIndexP}
\title{Estimates the index P by simulation using a theoretical climate range.
That is, the estimated posteriors of factors Eta and Rho are sampled,
as well as the distribution of factor Alpha. These independent samples are then
used to calculate index P given its mathematical expression. Since this
expression is dependent on climatic variables in time, the result is a time
series of the index P. The user can choose the number of samples,
resulting in \code{nSample} estimations of P in time. These independent
estimations are then used to calculate the mean and 95\% CI of the index P.}
\usage{
simulateTheoreticalIndexP(nSample = 1000)
}
\arguments{
\item{nSample}{Number of samples to draw from factors Alpha, Eta and Rho.}
}
\description{
In contrast to the function \code{simulateEmpiricalIndexP} which uses the
observed climatic variables, this function uses a theoretical range for climate.
The range can be set in the function's parameters.
}
\details{
The function adds the resulting index P estimations to the
\code{MVSE_results} global variable which can be used by other functions.
All independent estimations of index P in time are also stored3
in the \code{MVSE_indexP_theoretical} global variable (matrix, rows as independent estimations,
cols as time steps). To see a list of globally accessible variables use \code{help_globalVariables()}

The index P follows the expression of R0 from Lourenco et al. doi.org/10.7554/eLife.29820.001
ignoring the number of mosquitoes per human (V/N).
}
