% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{suitableSeason}
\alias{suitableSeason}
\title{Choose samples from the posterior and find the largest index P values.}
\usage{
suitableSeason(timethreshold = 0, Pthreshold = 1)
}
\arguments{
\item{timethreshold}{A non-negative integer value indicating the maximum time units P can drop below 1
while the interval of the 'suitable season' of P>=Pthreshold remains continuous.}

\item{Pthreshold}{A numeric value determining the threshold of P wanted to determine a suuitability season.}
}
\value{
Returns a data frame with years as columns, time as rows, and 1 indicating  P>=Pthreshold .
       This data frame is also saved as a csv file. Additionally a pdf containing a plot of the peak times is produced.
}
\description{
Choose samples from the posterior and find the largest index P values.
}
