% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSE.R
\name{weeklyAverage}
\alias{weeklyAverage}
\title{Transform daily date into averaged weekly values.}
\usage{
weeklyAverage(indexP, dates, nSamples = dim(indexP)[1])
}
\arguments{
\item{indexP}{A numeric matrix of Simulated index P, with rows as samples.}

\item{dates}{A character vector containing dates in a yyyy-mm-dd format. Should have the same length
as the number of columns in indexP.}
}
\value{
Returns a list with the the averaged index P as the first element and a vector of the corresponding
        dates as the second element. The data is broken by years so that every year starts at
        the same data as the first input date. This causes a loss of 1-2 days per year.
}
\description{
Transform daily date into averaged weekly values.
}
