% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climateFSA}
\alias{climateFSA}
\title{Example climate data}
\format{
A data frame with 468 rows 
\describe{
         \item{date}{Date in "\%Y-\%m-\%d" format}
         \item{T}{Temperature in degrees Celcius}
         \item{H}{Relative humidity}}
}
\usage{
climateFSA
}
\description{
Daily temperature (degrees Celcius) and relative humidity (%) data from 01-01-2015 in 01-01-2017 
in the city of Feira de Santana in Bahia, Brazil.
}
\keyword{datasets}
