% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvsefit.R
\name{extract}
\alias{extract}
\alias{extract,mvsefit-method}
\title{Extract samples from a fitted MVSE model}
\usage{
\S4method{extract}{mvsefit}(object, pars = NULL)
}
\arguments{
\item{pars}{A character vector of parameter names. The default is all parameters
(i.e. \code{"rho"}, \code{"eta"}, \code{"alpha"}, \code{"indexP"}, \code{"Q"}, \code{"V0"}).}

\item{x}{An object of class \code{\linkS4class{mvsefit}}.}
}
\value{
A named list, every element of which is either a numeric vector or dataframe representing samples
for a parameter.
}
\description{
Extract samples from a fitted model represented by an instance of class \code{\linkS4class{mvsefit}}.
}
